requirements.txt
Flask
numpy
pandas
tensorflow-cpu
werkzeug
scipy
pillow
requests
Flask-APScheduler